Redesign the chat input area
Moves to a new design for the input panel, renaming it to ChatInput. This change improves the user interface and experience by implementing a
new design with enhanced features such as attachment handling and
deep think. The old design is preserved in a separate file for now.
